chore(formula-vba): clippy cleanups + fix v3 transcript test
- Address clippy warnings in formula-vba (contains/is_multiple_of/repeat_n, lifetime elision, doc list indentation, minor refactors).

- Fix v3 content-normalization integration test to call project_normalized_data_v3_transcript (the full v3 ProjectNormalizedData transcript used for ContentsHashV3 binding) instead of the dir-record-only helper.

This keeps behavior the same while making cargo clippy -p formula-vba -- -D warnings and cargo test -p formula-vba pass.

+++ crates/formula-vba/src/compression.rs
@@ -261,7 +261,7 @@ fn copy_token_bit_count(current_decompressed_len: usize) -> u32 {
     let bits_needed = if n == 0 {
         0
     } else {
-        (usize::BITS - n.leading_zeros()) as u32
+        usize::BITS - n.leading_zeros()
     };
     bits_needed.clamp(4, 12)
 }
